LayoutTests/fast/dom/document-set-body-null.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<iframe id="testFrame" srcdoc="<html><body></body></html>"></iframe>
<script>
description("Tests setting document.body to null");

var testDocument = document.getElementById("testFrame").contentDocument;
shouldThrow("testDocument.body = null", "'HierarchyRequestError (DOM Exception 3): The operation would yield an incorrect node tree.'");

</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
